1 api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==4.4.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.13
decorator==5.1.1
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.112.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.29
langchain-text-splitters==0.2.2
langsmith==0.1.99
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
orjson==3.10.7
packaging==24.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.24.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.38.2
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.30.6
validators==0.20.0
watchdog==4.0.2
yarl==1.9.2
zipp==3.16.2
